Simplified  RemoveKeyDates Method.
Added Checkbox binding mode=TwoWay to better show intention

Changed the planner year view model to store a reference to the KeyDatesWindowViewModel so that it is not newed up and disposed of each time the window is opened.

Got the Calendar view highlighting all key dates correctly.

Added warning text to datepicker when adding a key date when user is selecting weekend date.